screams in visual studio fuckery
due to some weird fuckery, the c++20 conversion wasn't actually included in the last commit
(ps: explicit casts seem to cause vs to not complain about type changes)

+++ configdialog.cpp
@@ -112,7 +112,7 @@ void ConfigDialog::refresh() {
 	}
 
 	// these should probably be put somewhere else at some point
-	if (m_current_config.at(YonkPattern) < 0 || m_current_config.at(YonkPattern) >= _PATTERN_COUNT)
+	if (m_current_config.at(YonkPattern) < 0 || m_current_config.at(YonkPattern) >= (int)_PATTERN_COUNT)
 	{
 		m_current_config.at(YonkPattern) = 0;
 	}
